Update scala-for-java-programmers.md
Saying that Scala's `Any` is more general than Java's `Object` is not right as both are supertypes of every class. In Java not every data type is an object, there are also primitives, such as `int`, `long` and so on, but their existence does not make `Object` less general. Every primitive has a wrapper (`Integer`, `Long`, ...) extending in some way `Object` (through `Number`).